DSC_Jacard_Hausdorff/requirements.txt
filelock==3.15.4
fsspec==2024.6.0
imageio==2.34.2
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
tifffile==2024.6.18
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
